LUCENE-5078: Allow TfIDFSimilarity implementations to encode norm values into more than a single byte
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1496588 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/test/org/apache/solr/update/DocumentBuilderTest.java
@@ -357,9 +357,7 @@ public class DocumentBuilderTest extends SolrTestCaseJ4 {
    */
   private static byte expectedNorm(final DefaultSimilarity sim,
                                    final int length, final float boost) {
-    
-    return sim.encodeNormValue(boost / ((float) Math.sqrt(length)));
-
+    return (byte) sim.encodeNormValue(boost / ((float) Math.sqrt(length)));
   }
